requirements.txt
python-dotenv==1.2.1
Monzo-API==1.2.2
beautifulsoup4==4.14.2
requests==2.32.5
pandas==2.3.3
rich==14.2.0
openpyxl==3.1.5
